Fix packer build command format. Undo changes to FormatArgs.
